move combobox into own file
